+++ libindy/src/services/metrics/models.rs
@@ -3,12 +3,12 @@ use std::collections::HashMap;
 
 #[derive(Serialize, Deserialize)]
 pub struct MetricsValue {
-    value: usize,
+    value: u128,
     tags: HashMap<String, String>,
 }
 
 impl MetricsValue {
-    pub fn new(value: usize, tags: HashMap<String, String>) -> Self {
+    pub fn new(value: u128, tags: HashMap<String, String>) -> Self {
         MetricsValue { value, tags }
     }
 }
